requirements.txt
python-telegram-bot==20.7
anthropic==0.18.1